requirements.txt
python-dotenv==1.0.0
streamlit==1.33.0
langchain==0.2.0
langchain-community==0.2.4
PyGithub==1.55
pymongo
plotly 
pandas

